updated the api endpoints to work with userids as well as phone number lookups
