+++ server/api.js
@@ -1,11 +1,14 @@
 require('dotenv').config();
+require('./database');
 const path = require('path');
 const express = require('express');
 const api = express();
-const phoneVerification = require('./phoneVerification');
 const bodyParser = require('body-parser');
 const {logger} = require('./logger');
-const dynamoDB = require('./dynamoDB');
+const phoneVerificationController = require('./controllers/phoneVerification');
+const usersController = require('./controllers/users');
+const refreshTokensController = require('./controllers/refreshTokens');
+
 
 //passport stuff
 const passport = require('passport');
@@ -17,38 +20,38 @@ api.use(bodyParser.urlencoded({ extended: true }));   // to support URL-encoded 
 
 api.post('/request-verification', (req, res) => {
   logger.info('ExplorerAPI received request to /request-verification route');
-  phoneVerification.requestPhoneVerification(req, res);
+  phoneVerificationController.requestPhoneVerification(req, res);
 });
 
 api.post('/verify-token', (req, res) => {
   logger.info('ExplorerAPI received request to /verify route');
-  phoneVerification.verifyPhoneToken(req, res);
+  phoneVerificationController.verifyPhoneToken(req, res);
 });
 
 api.post('/search', function (req, res) {
   logger.info('Searching Users with phone details:');
-  dynamoDB.findUserWithPhone(req, res);
+  usersController.findUserWithPhone(req, res);
 });
 
 api.post('/settings', function (req, res) {
   logger.info('Updating user with received settings:');
-  dynamoDB.updateUserSettings(req, res);
+  usersController.updateUserSettings(req, res);
 });
 
 api.post('/token', function (req, res, next){
   logger.info('Checking the validity of the provided refresh token:');
-  dynamoDB.checkRefreshToken(req, res, next);
+  refreshTokensController.checkRefreshToken(req, res, next);
 });
 
 api.post('/token/reject', function (req, res, next) {
   logger.info('Revoking the provided refresh token:');
-  dynamoDB.revokeRefreshToken(req, res, next);
+  refreshTokensController.revokeRefreshToken(req, res, next);
 });
 
 // This endpoint is for dev purposes
 api.post('/users', (req, res) => {
   logger.info('ExplorerAPI received request to create draft user');
-  dynamoDB.createDraftUser(req, res);
+  usersController.createDraftUser(req, res);
 });
 
 // This endpoint is for dev purposes
