+++ server/controllers/phoneVerification.js
@@ -1,7 +1,6 @@
-const phoneReg = require('./lib/phoneVerification/phoneVerification')(process.env.TWILIO_API_KEY);
-const dynamoDB = require('./dynamoDB');
-const {logger} = require('./logger');
-
+const phoneReg = require('./../lib/phoneVerification/phoneVerification')(process.env.TWILIO_API_KEY);
+const dynamoDB = require('./../database');
+const {logger} = require('./../logger');
 
 /**
  * Register a phone
